src/requirements.txt
openai
requests
python-dotenv
pymongo
loguru